requirements.txt
tensorflow==2.15.0
numpy
matplotlib
pillow
